+++ src/index.ts
@@ -46,6 +46,8 @@ export * from "./core/InputCheckbox";
 export { default as InputCheckbox } from "./core/InputCheckbox";
 export * from "./core/InputDropdown";
 export { default as InputDropdown } from "./core/InputDropdown";
+export * from "./core/InputRadio";
+export { default as InputRadio } from "./core/InputRadio";
 export * from "./core/InputSearch";
 export { default as InputSearch } from "./core/InputSearch";
 export * from "./core/InputSlider";
@@ -72,8 +74,8 @@ export * from "./core/Notification";
 export { default as Notification } from "./core/Notification";
 export * from "./core/Pagination";
 export { default as Pagination } from "./core/Pagination";
-export * from "./core/Radio";
-export { default as RadioButton } from "./core/Radio";
+export * from "./core/InputRadio";
+export { default as RadioButton } from "./core/InputRadio";
 export * from "./core/SegmentedControl";
 export { default as SegmentedControl } from "./core/SegmentedControl";
 export * from "./core/styles";
